Refactor SurveyEngineCore and update tests
- Moved the SurveyEngineCore class to a new file for better organization.
- Updated import paths in test files to reflect the new location of SurveyEngineCore.
- Enhanced response handling tests to ensure correct behavior with prefills.
- Added a clone method to the ResponseItem class for improved functionality.
