+++ src/data_types/index.ts
@@ -1,3 +1,2 @@
 export * from './context';
-export * from '../survey/responses/response';
 export * from './legacy-types';
